+++ sdk/face/azure-ai-vision-face/src/main/java/com/azure/ai/vision/face/FaceSessionAsyncClient.java
@@ -6,7 +6,7 @@ package com.azure.ai.vision.face;
 import com.azure.ai.vision.face.implementation.FaceSessionClientImpl;
 import com.azure.ai.vision.face.implementation.MultipartFormDataHelper;
 import com.azure.ai.vision.face.implementation.models.DetectFromSessionImageRequest;
-import com.azure.ai.vision.face.models.CreateLivenessSessionContent;
+import com.azure.ai.vision.face.models.CreateLivenessSessionOptions;
 import com.azure.ai.vision.face.models.CreateLivenessWithVerifySessionContent;
 import com.azure.ai.vision.face.models.DetectFromSessionImageOptions;
 import com.azure.ai.vision.face.models.FaceAttributeType;
@@ -113,7 +113,7 @@ public final class FaceSessionAsyncClient {
      * }
      * </pre>
      *
-     * @param body Body parameter.
+     * @param options Options parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the request is rejected by server.
      * @throws ClientAuthenticationException thrown if the request is rejected by server on status code 401.
@@ -123,9 +123,9 @@ public final class FaceSessionAsyncClient {
      */
     @Generated
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Response<BinaryData>> createLivenessSessionWithResponse(BinaryData body,
+    public Mono<Response<BinaryData>> createLivenessSessionWithResponse(BinaryData options,
         RequestOptions requestOptions) {
-        return this.serviceClient.createLivenessSessionWithResponseAsync(body, requestOptions);
+        return this.serviceClient.createLivenessSessionWithResponseAsync(options, requestOptions);
     }
 
     /**
@@ -568,30 +568,6 @@ public final class FaceSessionAsyncClient {
     }
 
     /**
-     * Create a new detect liveness session.
-     *
-     * Please refer to https://learn.microsoft.com/rest/api/face/liveness-session-operations/create-liveness-session for
-     * more details.
-     *
-     * @param body Body parameter.
-     * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws HttpResponseException thrown if the request is rejected by server.
-     * @throws ClientAuthenticationException thrown if the request is rejected by server on status code 401.
-     * @throws ResourceNotFoundException thrown if the request is rejected by server on status code 404.
-     * @throws ResourceModifiedException thrown if the request is rejected by server on status code 409.
-     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     * @return session result of detect liveness on successful completion of {@link Mono}.
-     */
-    @Generated
-    @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<LivenessSession> createLivenessSession(CreateLivenessSessionContent body) {
-        // Generated convenience method for createLivenessSessionWithResponse
-        RequestOptions requestOptions = new RequestOptions();
-        return createLivenessSessionWithResponse(BinaryData.fromObject(body), requestOptions).flatMap(FluxUtil::toMono)
-            .map(protocolMethodData -> protocolMethodData.toObject(LivenessSession.class));
-    }
-
-    /**
      * Delete all session related information for matching the specified session id.
      *
      * Please refer to https://learn.microsoft.com/rest/api/face/liveness-session-operations/delete-liveness-session for
@@ -826,4 +802,29 @@ public final class FaceSessionAsyncClient {
     private static final TypeReference<List<FaceDetectionResult>> TYPE_REFERENCE_LIST_FACE_DETECTION_RESULT
         = new TypeReference<List<FaceDetectionResult>>() {
         };
+
+    /**
+     * Create a new detect liveness session.
+     *
+     * Please refer to https://learn.microsoft.com/rest/api/face/liveness-session-operations/create-liveness-session for
+     * more details.
+     *
+     * @param options Options parameter.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws HttpResponseException thrown if the request is rejected by server.
+     * @throws ClientAuthenticationException thrown if the request is rejected by server on status code 401.
+     * @throws ResourceNotFoundException thrown if the request is rejected by server on status code 404.
+     * @throws ResourceModifiedException thrown if the request is rejected by server on status code 409.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return session result of detect liveness on successful completion of {@link Mono}.
+     */
+    @Generated
+    @ServiceMethod(returns = ReturnType.SINGLE)
+    public Mono<LivenessSession> createLivenessSession(CreateLivenessSessionOptions options) {
+        // Generated convenience method for createLivenessSessionWithResponse
+        RequestOptions requestOptions = new RequestOptions();
+        return createLivenessSessionWithResponse(BinaryData.fromObject(options), requestOptions)
+            .flatMap(FluxUtil::toMono)
+            .map(protocolMethodData -> protocolMethodData.toObject(LivenessSession.class));
+    }
 }