+++ sdk/face/azure-ai-vision-face/src/main/java/com/azure/ai/vision/face/implementation/FaceSessionClientImpl.java
@@ -160,7 +160,7 @@ public final class FaceSessionClientImpl {
         @UnexpectedResponseExceptionType(HttpResponseException.class)
         Mono<Response<BinaryData>> createLivenessSession(@HostParam("endpoint") String endpoint,
             @HostParam("apiVersion") String apiVersion, @HeaderParam("Content-Type") String contentType,
-            @HeaderParam("Accept") String accept, @BodyParam("application/json") BinaryData body,
+            @HeaderParam("Accept") String accept, @BodyParam("application/json") BinaryData options,
             RequestOptions requestOptions, Context context);
 
         @Post("/detectLiveness-sessions")
@@ -171,7 +171,7 @@ public final class FaceSessionClientImpl {
         @UnexpectedResponseExceptionType(HttpResponseException.class)
         Response<BinaryData> createLivenessSessionSync(@HostParam("endpoint") String endpoint,
             @HostParam("apiVersion") String apiVersion, @HeaderParam("Content-Type") String contentType,
-            @HeaderParam("Accept") String accept, @BodyParam("application/json") BinaryData body,
+            @HeaderParam("Accept") String accept, @BodyParam("application/json") BinaryData options,
             RequestOptions requestOptions, Context context);
 
         @Delete("/detectLiveness-sessions/{sessionId}")
@@ -384,7 +384,7 @@ public final class FaceSessionClientImpl {
      * }
      * </pre>
      * 
-     * @param body Body parameter.
+     * @param options Options parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the request is rejected by server.
      * @throws ClientAuthenticationException thrown if the request is rejected by server on status code 401.
@@ -393,12 +393,12 @@ public final class FaceSessionClientImpl {
      * @return session result of detect liveness along with {@link Response} on successful completion of {@link Mono}.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Response<BinaryData>> createLivenessSessionWithResponseAsync(BinaryData body,
+    public Mono<Response<BinaryData>> createLivenessSessionWithResponseAsync(BinaryData options,
         RequestOptions requestOptions) {
         final String contentType = "application/json";
         final String accept = "application/json";
         return FluxUtil.withContext(context -> service.createLivenessSession(this.getEndpoint(),
-            this.getServiceVersion().getVersion(), contentType, accept, body, requestOptions, context));
+            this.getServiceVersion().getVersion(), contentType, accept, options, requestOptions, context));
     }
 
     /**
@@ -462,7 +462,7 @@ public final class FaceSessionClientImpl {
      * }
      * </pre>
      * 
-     * @param body Body parameter.
+     * @param options Options parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the request is rejected by server.
      * @throws ClientAuthenticationException thrown if the request is rejected by server on status code 401.
@@ -471,11 +471,11 @@ public final class FaceSessionClientImpl {
      * @return session result of detect liveness along with {@link Response}.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Response<BinaryData> createLivenessSessionWithResponse(BinaryData body, RequestOptions requestOptions) {
+    public Response<BinaryData> createLivenessSessionWithResponse(BinaryData options, RequestOptions requestOptions) {
         final String contentType = "application/json";
         final String accept = "application/json";
         return service.createLivenessSessionSync(this.getEndpoint(), this.getServiceVersion().getVersion(), contentType,
-            accept, body, requestOptions, Context.NONE);
+            accept, options, requestOptions, Context.NONE);
     }
 
     /**