+++ sdk/face/azure-ai-vision-face/src/test/java/com/azure/ai/vision/face/tests/LivenessSessionTest.java
@@ -6,7 +6,7 @@ package com.azure.ai.vision.face.tests;
 import com.azure.ai.vision.face.FaceServiceVersion;
 import com.azure.ai.vision.face.FaceSessionAsyncClient;
 import com.azure.ai.vision.face.FaceSessionClient;
-import com.azure.ai.vision.face.models.CreateLivenessSessionContent;
+import com.azure.ai.vision.face.models.CreateLivenessSessionOptions;
 import com.azure.ai.vision.face.models.LivenessOperationMode;
 import com.azure.ai.vision.face.models.LivenessSession;
 import com.azure.ai.vision.face.models.OperationState;
@@ -41,8 +41,8 @@ public class LivenessSessionTest extends FaceClientTestBase {
     public void testCreateSession(String httpClientName, FaceServiceVersion serviceVersion,
         Supplier<ILivenessSessionSyncCommands> commandProvider) {
         String uuid = UUID.randomUUID().toString();
-        CreateLivenessSessionContent content
-            = new CreateLivenessSessionContent(LivenessOperationMode.PASSIVE).setDeviceCorrelationId(uuid);
+        CreateLivenessSessionOptions content
+            = new CreateLivenessSessionOptions(LivenessOperationMode.PASSIVE).setDeviceCorrelationId(uuid);
         createSessionAndVerify(commandProvider.get(), content);
     }
 
@@ -50,8 +50,8 @@ public class LivenessSessionTest extends FaceClientTestBase {
     @MethodSource("getTestCommands")
     public void testCreateSessionDeviceIdOptional(String httpClientName, FaceServiceVersion serviceVersion,
         Supplier<ILivenessSessionSyncCommands> commandProvider) {
-        CreateLivenessSessionContent content
-            = new CreateLivenessSessionContent(LivenessOperationMode.PASSIVE).setDeviceCorrelationIdSetInClient(true);
+        CreateLivenessSessionOptions content
+            = new CreateLivenessSessionOptions(LivenessOperationMode.PASSIVE).setDeviceCorrelationIdSetInClient(true);
 
         createSessionAndVerify(commandProvider.get(), content);
     }
@@ -64,8 +64,8 @@ public class LivenessSessionTest extends FaceClientTestBase {
 
         int authTokenTimeToLiveInSeconds = 60;
         String uuid = UUID.randomUUID().toString();
-        CreateLivenessSessionContent content
-            = new CreateLivenessSessionContent(LivenessOperationMode.PASSIVE).setDeviceCorrelationId(uuid)
+        CreateLivenessSessionOptions content
+            = new CreateLivenessSessionOptions(LivenessOperationMode.PASSIVE).setDeviceCorrelationId(uuid)
                 .setAuthTokenTimeToLiveInSeconds(authTokenTimeToLiveInSeconds);
 
         createSessionAndVerify(livenessCommands, content);
@@ -95,7 +95,7 @@ public class LivenessSessionTest extends FaceClientTestBase {
     }
 
     private LivenessSession createSessionAndVerify(ILivenessSessionSyncCommands livenessCommands,
-        CreateLivenessSessionContent content) {
+        CreateLivenessSessionOptions content) {
         LivenessSession result = livenessCommands.createLivenessSessionSync(content);
 
         mCurrentCommand = livenessCommands;