+++ sdk/face/azure-ai-vision-face/src/test/java/com/azure/ai/vision/face/tests/commands/liveness/LivenessSessionSyncCommands.java
@@ -4,7 +4,7 @@
 package com.azure.ai.vision.face.tests.commands.liveness;
 
 import com.azure.ai.vision.face.FaceSessionClient;
-import com.azure.ai.vision.face.models.CreateLivenessSessionContent;
+import com.azure.ai.vision.face.models.CreateLivenessSessionOptions;
 import com.azure.ai.vision.face.models.LivenessSession;
 
 class LivenessSessionSyncCommands implements ILivenessSessionSyncCommands {
@@ -14,7 +14,7 @@ class LivenessSessionSyncCommands implements ILivenessSessionSyncCommands {
         this.mSyncClient = faceSessionClient;
     }
 
-    public LivenessSession createLivenessSessionSync(CreateLivenessSessionContent content) {
+    public LivenessSession createLivenessSessionSync(CreateLivenessSessionOptions content) {
         return mSyncClient.createLivenessSession(content);
     }